+++ cadl-tests/src/main/java/com/outputbasic/models/OutputModel.java
@@ -21,7 +21,7 @@ public final class OutputModel {
      * Required int, illustrating a value type property.
      */
     @JsonProperty(value = "requiredInt", required = true)
-    private long requiredInt;
+    private int requiredInt;
 
     /**
      * Creates an instance of OutputModel class.
@@ -32,7 +32,7 @@ public final class OutputModel {
     @JsonCreator
     public OutputModel(
             @JsonProperty(value = "requiredString", required = true) String requiredString,
-            @JsonProperty(value = "requiredInt", required = true) long requiredInt) {
+            @JsonProperty(value = "requiredInt", required = true) int requiredInt) {
         this.requiredString = requiredString;
         this.requiredInt = requiredInt;
     }
@@ -51,7 +51,7 @@ public final class OutputModel {
      *
      * @return the requiredInt value.
      */
-    public long getRequiredInt() {
+    public int getRequiredInt() {
         return this.requiredInt;
     }
 }
